Tied apt select to a specific commit
Tidied up the overall layout of the Dockerfile as it made my eyes hurt. Also, by using master.tar.gz without a check there was a potential risk of the content changing without our knowledge. I therefore propose we modify this git commit every month to coincide with the release of the baseimage going forward.